Wordle Bot
This is a program made to solve the wordle puzzle. It has a monte carlo simulator to test accuracy, a system to find the best word for every scenario, as well as a test for individual words. This program has a few bugs that need to be fixed as well as some optimizations to the algorithm to find the best word